move deathlink option tests to TestOptions.py
